Implementació de l'INSERT a base de dades

+++ src/Contact/Contact.java
@@ -3,10 +3,10 @@ package Contact;
 public class Contact {
     public String nom;
     public String cognom;
-    public int telefon;
+    public long telefon;
     public String direccio;
 
-    public Contact(String nom,String cognom,int telefon,String direccio) {
+    public Contact(String nom,String cognom,long telefon,String direccio) {
         this.nom = nom;
         this.cognom = cognom;
         this.telefon = telefon;
